pyproject.toml
[project]
name = "ai-learning-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.203.1",
    "crewai-tools>=0.76.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]
